+++ validator-core/src/main/java/no/difi/vefa/validator/api/RendererInfo.java
@@ -10,7 +10,7 @@ import java.lang.annotation.Target;
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.TYPE)
-public @interface PresenterInfo {
+public @interface RendererInfo {
 
     /**
      * List of file extensions.