+++ validator-core/src/main/java/no/difi/vefa/validator/presenter/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Implementations used for presentation.
- */
-package no.difi.vefa.validator.presenter;