+++ validator-core/src/main/java/no/difi/vefa/validator/renderer/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Implementations used for rendering.
+ */
+package no.difi.vefa.validator.renderer;